chore: move from helpers to utils folder
there were some files in the helpers folder, they should be in the utils
folder
